Make some loops in the SchurPreconditioner parallel
To do this, we also had to change the orientation of the Sk matrix
from the SchurComplement, because the Column major version of
SumIntoGlobalValues is not threadsafe and also slower. Because we
are still not sure if the Row major version is threadsafe, we left
the critical section around the call. This doesn't seem to slow
things down.

git-svn-id: https://vcs.service.rug.nl/svn/fredwubs/hymls@909 5bcbc087-4b5e-4533-96d0-4097f40ffae9