Remove the find files for cmake, because phist already has config files
They can be found if the install dir has the bin path added to the PATH.
So if I install phist in ~/local, PATH should be ${PATH}:${HOME}/local/bin.

This currectly doesn't work yet, because phist installs to the wrong path,
but as soon as Melven fixes this it will work.

I did this change now because Melven changed the package name from PHIST
to phist, so the old cmake files didn't work anymore

git-svn-id: https://vcs.service.rug.nl/svn/fredwubs/hymls@1195 5bcbc087-4b5e-4533-96d0-4097f40ffae9