Disabled a workaround for the 3D Stokes Hybrid stuff if there are not 4 dof.
In the future I hope we get around all the StokesClassifier stuff by using a sparter
partitioning that doesn't suffer from full conservation cells and tubes (FCCs and FCTs).
This commit basically fixes an issue for the 3D scalar case with periodici BC, which we
need for the Turing problem.