Refactor the SchurComplement and Preconditioner code related to matrix blocks A_{[12],[12]}
I added a MatrixBlock class which contains these blocks. Reason I do this is so the
Preconditioner code looks more clean, and we don't have to pass a mother argument to
the SchurComplement. I think this mother argument causes a circular reference
between Preconditioner and SchurComplement, meaning that neither of both will ever
be deleted.

At the moment this mother argument is still needed, because some things like the
ApplyInverse for the A11 block remain unimplemented. After I finish this, we can
remove the mother argument, and also remove all duplicate work that is done in both
the Preconditioner and MatrixBlock instances at the moment.