Disable all BorderedLU stuff in the Solver
We now also use the Belos preconditioning interface and not the one
in the ProjectedOperator. This also seems to allow for preconditioning
from both sides.

One bug here was that if the BorderedLU stuff was enabled in the Solver
always an initial guess Zero was used. As soon as I disabled this a
random initial guess was used, which broke things, because it was not
orthogonalized beforehand.

Refs #12 #14 #15