Add a SetBorder method to the Solver class that we can use in JDQR
Now we can set the two borders separately, which we needed for JDQR
because on was BV and the other V. Now we can also use both JDQR and
a null space at the same time. You first call setNullSpace and then
SetBorder in JDQR. This will add the border from JDQR to the null space.

Refs #14 #15