Re #13, added some \todo's in HYMLS_Solver.H and tried to improve the documentation a little
Right now it seems that, in order to add the null space to the preconditioner, we have to
first call Solver::setNullSpace, then (optionally) Solver::SetBorder (for Jacobi-Davidson), and finally
Solver::SetupDeflation.

When NOX uses our Solver, the interface is NOX::Epetra::LinearSystemHymls. There, SetupDefalation is called,
which is why it worked for Weiyan before. In main.C we only call setNullSpace right now, so it doesn't work
(why does it work for the neumann.xml test case???)