Reimplement Preconditioner ApplyInverse
To prevert excessive imports, we had to reimplement the matrix
blocks on top of the original matrix, not the extended reordered
matrix that we originally used. Ideal would be to get rid
of this reordered matrix.

We also made the application of the A11 solve more generic by
converting between maps while setting the LHS and RHS