Re #47, in the phist interface we now set the approximate nullspace on the preconditioner as
P        P\V
(BV)'   (BV)'(P\V)

the option "Bordered Solver" in subspacejada now seems to work for the Laplace problem,
for non-symmetric cases and with B!=I we have to verify this.

The idea here is to mimic the "skew projection" of JDQR, the previous border

P       V
(BV)'   0

mimics a regular orthogonal projection instead. The previous implementation using SetupDeflation
probably worked because the BorderedLU class uses the same borders as above, except it solves more
accurately for A\V using GMRES (this remark fixes #32, I guess).

see also #13, #14, #30.