added a sandbox directory for experimenting with things that haven't entered the HYMLS code.
As a start, there's a skew3D folder with a matlab script to visualize 'skew separators' in 3D,
right now it just plots cell indices in octree ordering and draws some lines and surfaces depending on which flags are
set in the script (skew3D_plot.m). The cube splitting it draws is *not* what we want in the end.