Always compute the next level border in Compute instead of setBorder.
This makes it so that you only have to compute the factorization once
as long as you call setBorder before Compute. We now throw a warning
if you don't do this.

This also has the advantage of only having to call setBorder once,
instead of after every Compute, but this actually caused problems
in the BorderedDeflatedSolver. This may need further investigation.